新增点赞,收藏接口

+++ _internal/article/_internal/repository/dao/entity.go
@@ -1,11 +1,11 @@
 package dao
 
 type Article struct {
-	Id       int
+	Id       int64
 	Title    string
 	Content  string
 	Status   uint
-	AuthorId int
+	AuthorId int64
 	Ctime    int64
 	Utime    int64
 }
